Added very insecure but simple encryption
Added a very insecure but simple encryption. Use a password to encrypt
any object into a CryptedObject and decrypt it using the same password
to get your object back. That is a symmetric encryption based on some
simple math and can be cracked with some know how from Wikipedia, but
might be enough for the simplest applications. A way to use this might
be a remote/web server broadcasting a new password to random times the
clients can request via a secure SSL connection or something like that.